+++ aidungeon/interface/colors-classic.ini
