+++ aidungeon/interface/colors-full.ini
