--- changelog.txt
@@ -1,19 +1,0 @@
-removed botnet (uploading saved stories to public google drive)
-removed useless profanity filter
-added config file
-added color
-added download of user generated prompts
-added prompt selection menu
-added console bell alert when AI finishes generating
-removed useless dependencies
-added text wrapping
-added ability to save custom prompts
-added color support on windows via colorama
-changed to use pytorch gpt using the transformers module
-added ability to use 16bit mode on GPU to save significant VRAM
-added suggested actions
-added "d20" randomization of action sucess/failure
-added repetitiveness penalty for AI
-added ability to select from multiple models
-added simple to use script to convert tensorflow models
-probably a lot of stuff I forgot to add here...
