+++ config.ini
@@ -62,7 +62,7 @@ text-wrap-width = 120
 force-cpu = off
 
 # 30 will not spam you with console log message, <30 will spam devs
-log-level = 20
+log-level = 30
 
 # use a dice to decide actions success. E.g. rolling a 1 means "You failed to X"
 action-d20 = on
@@ -90,10 +90,10 @@ prompt-toolkit = on
 autosave = on
 
 #Color scheme that is used if Python Prompt Toolkit is available. A classic-type color scheme can still be used here.
-color-scheme = interface/colors-full.ini
+color-scheme = aidungeon/interface/colors-full.ini
 
 #Backup color scheme in case Python Prompt Toolkit isn't available
-backup-color-scheme = interface/colors-classic.ini
+backup-color-scheme = aidungeon/interface/colors-classic.ini
 
 #Use experimental gpt2 (may be slightly faster, but buggy)
 gpt2-experimental = off
