+++ play.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+venv/bin/python launch.py
